requirements.txt
Flask==0.8
Jinja2==2.6
Werkzeug==0.8.3
certifi==0.0.8
chardet==1.0.1
gunicorn==0.14.2
pysandbox==1.5
redis==2.4.11
requests==0.10.8
simplejson==2.4.0
wsgiref==0.1.2
